Add .nvmrc file
If the developer has `nvm` installed and it's configured properly, it will auto switch to the specified Node.js version when the terminal directory changes to the repo's root.

If `nvm` is not configured to auto-switch, we can just execute `nvm use` and it will switch to the specified Node.js version.